requirements.txt
streamlit
pandas
plotly
yfinance
google-generativeai
requests
# optional if you use specific versions
